refac guiText
